refactor: move null handling code
