+++ package/src/index.ts
@@ -2,7 +2,6 @@ import { transaction } from './operations/transaction'
 import { HybridNitroSQLite } from './nitro'
 import { open } from './operations/session'
 import { execute, executeAsync } from './operations/execute'
-import { SQLiteNullValue } from './types'
 import { init } from './OnLoad'
 export type * from './types'
 export { typeORMDriver } from './typeORM'
@@ -22,21 +21,9 @@ export const NitroSQLite = {
 }
 
 export { open } from './operations/session'
-
-let ENABLE_SIMPLE_NULL_HANDLING = false
-export function enableSimpleNullHandling(
-  shouldEnableSimpleNullHandling = true
-) {
-  ENABLE_SIMPLE_NULL_HANDLING = shouldEnableSimpleNullHandling
-}
-export function isSimpleNullHandlingEnabled() {
-  return ENABLE_SIMPLE_NULL_HANDLING
-}
-
-export const NITRO_SQLITE_NULL: SQLiteNullValue = { isNitroSQLiteNull: true }
-export function isNitroSQLiteNull(value: any): value is SQLiteNullValue {
-  if (typeof value === 'object' && 'isNitroSQLiteNull' in value) {
-    return true
-  }
-  return false
-}
+export {
+  isNitroSQLiteNull,
+  NITRO_SQLITE_NULL,
+  isSimpleNullHandlingEnabled,
+  enableSimpleNullHandling,
+} from './nullHandling'
